Editor: Refactor mouse_over_ui detection
The previous method used a Control node beneath the editor UI to detect when the mouse was over the control node (and not overlapping the editor UI). This didn't work on mobile devices, so I replaced it with a programmatic solution, where the mouse position is evaluated against every UI node in the editor using recursion.